Generify hand definition and separate code to files
To test things are working correctly, I'm going to write
some tests all types of hands. To that end, this generifies
what a hand definition is. So now there is a "Hand" which
is a realized collection of cards. And there is a "Hand Definition"
which is a special type of hand that has a probability of occurring:
pair, flush, straight, nothing, etc.